[Hasher Persistence] Hasher data model now persists its own mother kennel data
Problem:
Loading a hasher requires loading a kennel, at the logic level.

Reasoning:
A hasher can be editted if need be.  If their mother kennel changes name
or acronym, which is highly unlikley, they can always be edited or maybe
they want to keep the original values?  Who knows.

Implementation:
Mother kennel is now persisted as a MapAttribute that holds kennel_id,
name and acronym.
mother_kennel_name_lower is also persisted and is used in an index as a
range key.