[Kennel Membership persistence] Added hasher and kennel data to membership record
Problem: Using the KennelMembership table I could get IDs membership information
for hashers and kennels with IDs, but I still needed to query potentially
both sides of the membership to make this relationship displayable.

Resolution: Added hasher_Ref and kennel_ref to the membership record.
This data is specified by the Kennel or Hasher Reference model, and
those models can be altered to include more or less in the membership
record.