[Location] Added uniqueness check on save.
Change: A uniqueness check was needed to prevent the same location
 getting multiple entries.

Implementation:
Uniqueness is consider to be a location name that is lower cased and
stripped of whitespace (searchable_name) within a proximity of a geohash
block.  If the geo hash is AAAA1111, anything within AAAA1110 through
AAAA1120 will be consider 'in proximity'.

An index was added to perform lookups of this nature.

also added searchable address, and a lat/long index.
These may be useless...