[Location] Added SearchableNameIndex and searchable name prefix and suffix to persistence model.
Need: A mechanism by which to search for a part of a location name and provide suggestions.

Implementation: searchable name is split into a prefix of three characters, used
as a hash key, and a suffix of the remaining string, used as a range key.

This can be used to return a list of partially matching names.